### 2025-12-06 #9568 @editorialbot
[PRE REVIEW]: Nexarag: Democratizing Reproducible Knowledge Graph Contexts for LLM Research
**Submitting author:** @tjkerby (<a href="http://orcid.org/0009-0002-1189-3820">Thomas Kerby</a>)
**Repository:** https://github.com/KevinMoonLab/Nexarag
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
