### 2026-01-16 #9568 @crvernon
👋 @yewentao256 - are you able to take this one on as editor?

@tjkerby will have to update the submission to the new JOSS requirements, but once this is done you can proceed with finding reviewers.
