### 2026-01-21 #9568 @yewentao256
Hi @tjkerby,

Just a gentle reminder to update the paper. Once that’s done, I can move forward with finding reviewers for you.

Thanks!
