### 2026-01-21 464c152
Expand documentation with new sections
Added sections on software design, use cases, quality control, research impact, and AI usage disclosure to enhance the documentation of Nexarag.
