+++ paper.md
@@ -32,9 +32,10 @@ Large language models (LLMs) are widely used in research workflows but struggle 
 
 Traditional retrieval‑augmented generation (RAG) systems rely primarily on embedding‑based similarity [@Lewis2020; @Guu2020; @reimers-2019-sentence-bert], which often misses long‑range semantic structure and relationships across documents, especially in long‑context, multi‑document settings [@wang-etal-2024-leave; @gao2023retrieval]. This weakens controllability, auditability, and reproducibility for complex research tasks. Knowledge graphs provide a structured and interpretable alternative by modeling entities and relations explicitly [@reinanda2020knowledge]. However, existing KG‑powered tooling is either proprietary and expensive or technically demanding to deploy and maintain. Nexarag fills this gap with a researcher‑friendly platform that automates KG creation from academic inputs, supports semantic exploration, and standardizes LLM access via MCP [@anthropic2024mcp], enabling reproducible literature synthesis across local and cloud settings.
 
-# State of the field
-
 RAG improves access to external knowledge and has become a standard strategy for knowledge‑intensive NLP [@Lewis2020; @Guu2020; @gao2023retrieval], yet it struggles with long contexts and multi‑step reasoning when similarity search is the only primitive [@wang-etal-2024-leave]. KGs address this by enabling path‑based queries and explicit relation reasoning while preserving transparency and updatability [@reinanda2020knowledge; @sahlab2022knowledge; @Xu2024]. Nexarag’s contribution is to operationalize these advantages in a package that researchers can run locally, share with collaborators, and connect to a wide range of LLM hosts through MCP [@anthropic2024mcp].
+
+# Software Design
+In designing Nexarag we emphasized four principles: (1) ease-of-use (in deployment and practical application), (2) flexibility (in model selection and configuration), (3) modularity (to promote scale and independent contribution), and (4) privacy and security. For ease-of-use, we chose popular frontend technologies like Angular, D3.js, and Cytoscape.js and leveraged existing component libraries to provide a simple and familiar frontend user interface with intuitive tools for building knowledge graphs, searching for relevant papers, conversing with LLMs, and visualizing data. Nexarag is also fully containerized, with release artifacts produced by automated build pipelines readily available for local deployment through Docker. For flexibility, we integrate with Ollama and support any embedding model and LLM that is also supported by the user's hardware, making it simple to switch between models for different tasks and as new models are released. Users can also plug in their preferred LLM or coding agent of choice using the built-in MCP server. Additionally, Nexarag features a highly modular design, with distinct services for the REST API, the neo4j knowledge graph, the MCP server, and the frontend application, all bound together with a RabbitMQ messaging backbone. This allows components to scale horizontally, and to minimize the blast radius of contributions in a single service. Finally, Nexarag supports on-premises, air-gapped deployments, offering privacy and security that is not available in cloud-based applications.
 
 # Software overview
 
@@ -65,20 +66,29 @@ ollama pull gemma3:1b
 
 **License:** GNU General Public License v3.0.
 
-# Quality control
-
-Nexarag emphasizes verifiable operation through containerized deployment and a guided quick start [@boettiger2015docker]. Reviewers can launch the full stack with Docker Compose, query/persist KGs in Neo4j, and exercise end‑to‑end flows (semantic search, citation expansion, MCP tools). A worked MCP chat transcript and an automatically generated literature review illustrate that the system’s graph building, retrieval, and reporting features execute as described. The repository includes example datasets/notebooks and scripts for running tests where applicable, supporting broader reproducibility goals in research practice [@rothacher2023eleven].
-
-# Use cases
+## Use cases
 
[…]
