### 2026-01-21 69726ea
Update AI usage disclosure section with specifics
Added details about AI tools used in development.
