+++ paper.md
@@ -85,7 +85,10 @@ Although Nexarag is a relatively new project and has not yet accumulated extensi
 Nexarag is positioned to serve as shared research infrastructure for studies on retrieval-augmented generation, knowledge-graph–augmented reasoning, and AI-assisted literature review workflows. Its model-agnostic design, enabled by MCP, allows researchers to interchange local or API-hosted LLMs while holding the underlying knowledge graph and retrieval logic fixed. This supports direct comparison of LLM behavior under identical, graph-derived contexts, facilitating methodological research on controllability, hallucination reduction, and long-context reasoning. By lowering the technical barrier to building, inspecting, and sharing reproducible knowledge graph contexts, Nexarag enables researchers to move beyond ad hoc, model-coupled RAG pipelines toward more transparent and portable AI-assisted research practices.
 
 # AI usage disclosure
-Generative AI tools were used in the development of the software, supporting code reviews, providing minor features in the frontend, and identifying and fixing bugs. It was also used to generate some of the documentation, and assisted with paper authoring.
+Generative AI tools were used in the development of the software, supporting code reviews, providing minor features in the frontend, and identifying and fixing bugs. It was also used to generate some of the documentation, and assisted with paper authoring. We primarily used:
+
+* ChatGPT with the GPT-4o model for writing tasks
+* Claude Code with the Sonnet 4 model for coding tasks
 
 # Acknowledgements
 
