### 2026-01-21 163e89d
Include review process for AI-generated content
Added a note about AI-generated material review and design decisions.
