### 2026-01-21 ef34755
Update AI usage disclosure section in paper.md
Clarified the role of generative AI tools in software development and documentation.
