### 2026-01-22 #9568 @tjkerby
Hi @yewentao256,

Thank you for the gentle reminder. We have made the necessary alterations to our paper.md file to conform to the new submission requirements.

Let us know if you need anything else on our end to continue the review process.
