### 2026-01-26 64d338d
Add state of the field section on RAG and KGs
Added section on the state of the field regarding retrieval-augmented generation and knowledge graph tools, highlighting existing frameworks and their limitations.
