+++ paper.md
@@ -34,6 +34,18 @@ Traditional retrieval‑augmented generation (RAG) systems rely primarily on emb
 
 RAG improves access to external knowledge and has become a standard strategy for knowledge‑intensive NLP [@Lewis2020; @Guu2020; @gao2023retrieval], yet it struggles with long contexts and multi‑step reasoning when similarity search is the only primitive [@wang-etal-2024-leave]. KGs address this by enabling path‑based queries and explicit relation reasoning while preserving transparency and updatability [@reinanda2020knowledge; @sahlab2022knowledge; @Xu2024]. Nexarag’s contribution is to operationalize these advantages in a package that researchers can run locally, share with collaborators, and connect to a wide range of LLM hosts through MCP [@anthropic2024mcp].
 
+# State of the field
+
+The open-source ecosystem around retrieval-augmented generation (RAG) has matured quickly, with widely used orchestration frameworks such as LangChain, LlamaIndex, and Haystack providing reusable components for ingesting documents, building indexes, and implementing retrieve-then-generate pipelines [@langchain2026rag; @llamaindex2026rag; @haystack2026]. These frameworks are well suited for building bespoke RAG applications, but they are primarily developer libraries: users typically assemble pipelines in code, and the resulting “context construction” logic is often tightly coupled to a specific project’s embeddings, chunking choices, runtime configuration, and LLM host.
+
+Graph-augmented retrieval approaches have emerged to address limitations of similarity-only retrieval for multi-document reasoning. For example, LlamaIndex provides a KnowledgeGraphIndex that supports automated knowledge graph construction from unstructured text and entity-based querying [@llamaindex2026kg]. LangChain offers graph question-answering utilities, including GraphCypherQAChain for translating natural-language questions into Cypher queries over Neo4j [@langchain2026neo4j]. Dedicated GraphRAG toolkits such as Microsoft GraphRAG and the Neo4j GraphRAG package for Python provide pipelines for extracting structured representations from text and using graph-aware retrieval strategies during LLM inference [@microsoft2026graphrag; @neo4j2026graphrag]. These projects demonstrate that knowledge graphs can improve retrieval structure and interpretability, but they are generally positioned as building blocks for developers rather than as reproducible, end-to-end research applications for literature-centric workflows.
+
+In parallel, literature discovery tools such as Connected Papers, ResearchRabbit, and Litmaps offer citation-network visualizations and recommendations to help researchers explore related work and track research areas over time [@connectedpapers2026; @researchrabbit2026; @litmaps2026]. While valuable for interactive discovery, these tools are not designed to serve as a researcher-owned, queryable knowledge graph substrate that can be versioned, shared, and integrated as a controllable context source for LLM experiments across local and secure environments.
+
+Nexarag fills the gap between these two tool families by packaging knowledge-graph-based context construction as a reproducible, shareable research artifact and workflow. Instead of treating graph retrieval as a code-level feature, Nexarag provides a self-hostable platform centered on a persistent Neo4j knowledge graph, interactive graph curation and visualization, and standardized LLM access through the Model Context Protocol (MCP) [@anthropic2024mcp; @mcp2024github]. This design enables researchers to hold the underlying graph context and retrieval tools fixed while varying models, prompts, and deployment settings, supporting more transparent and reproducible LLM-assisted literature synthesis.
+
[…]
