### 2026-01-26 1cb34f3
Add references for state of the field section.
Added multiple references related to RAG and associated tools.
