+++ paper.bib
@@ -298,3 +298,73 @@
   pages={2},
   year={2014}
 }
+
+@misc{langchain2026rag,
+  title        = {Build a RAG agent with LangChain},
+  author       = {{LangChain}},
+  howpublished = {\url{https://docs.langchain.com/oss/python/langchain/rag}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{llamaindex2026rag,
+  title        = {Introduction to RAG (Retrieval-Augmented Generation)},
+  author       = {{LlamaIndex}},
+  howpublished = {\url{https://developers.llamaindex.ai/python/framework/understanding/rag/}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{haystack2026,
+  title        = {Get Started (Haystack Documentation)},
+  author       = {{deepset Haystack}},
+  howpublished = {\url{https://docs.haystack.deepset.ai/docs/get-started}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{llamaindex2026kg,
+  title        = {Knowledge Graph Index (LlamaIndex Python Documentation)},
+  author       = {{LlamaIndex}},
+  howpublished = {\url{https://developers.llamaindex.ai/python/examples/index_structs/knowledge_graph/knowledgegraphdemo/}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{langchain2026neo4j,
+  title        = {Neo4j (LangChain Integration Documentation)},
+  author       = {{LangChain}},
+  howpublished = {\url{https://docs.langchain.com/oss/python/integrations/providers/neo4j}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{microsoft2026graphrag,
+  title        = {microsoft/graphrag: A modular graph-based Retrieval-Augmented Generation (RAG) system},
+  author       = {{Microsoft}},
+  howpublished = {\url{https://github.com/microsoft/graphrag}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{neo4j2026graphrag,
+  title        = {neo4j/neo4j-graphrag-python: Neo4j GraphRAG Package for Python},
+  author       = {{Neo4j}},
+  howpublished = {\url{https://github.com/neo4j/neo4j-graphrag-python}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{connectedpapers2026,
+  title        = {Connected Papers},
+  author       = {{Connected Papers}},
+  howpublished = {\url{https://www.connectedpapers.com/}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{researchrabbit2026,
+  title        = {ResearchRabbit},
+  author       = {{ResearchRabbit}},
+  howpublished = {\url{https://www.researchrabbit.ai/}},
+  note         = {Accessed: 2026-01-23}
+}
+
+@misc{litmaps2026,
+  title        = {Litmaps},
+  author       = {{Litmaps}},
+  howpublished = {\url{https://www.litmaps.com/}},
+  note         = {Accessed: 2026-01-23}
+}
