### 2026-01-26 c9ee899
Update paper.md
