+++ paper.md
@@ -44,7 +44,7 @@ In parallel, literature discovery tools such as Connected Papers, ResearchRabbit
 
 Nexarag fills the gap between these two tool families by packaging knowledge-graph-based context construction as a reproducible, shareable research artifact and workflow. Instead of treating graph retrieval as a code-level feature, Nexarag provides a self-hostable platform centered on a persistent Neo4j knowledge graph, interactive graph curation and visualization, and standardized LLM access through the Model Context Protocol (MCP) [@anthropic2024mcp; @mcp2024github]. This design enables researchers to hold the underlying graph context and retrieval tools fixed while varying models, prompts, and deployment settings, supporting more transparent and reproducible LLM-assisted literature synthesis.
 
-We chose to build Nexarag rather than contribute directly to an existing RAG or GraphRAG library because our core contribution requires application-level infrastructure that is outside the scope of most libraries’ design constraints: a literature-oriented ingestion workflow (BibTeX and paper lists, citation expansion, and persistent graph storage), a collaborative and visual curation interface, containerized offline deployment, and an MCP-first tool surface for model-agnostic experimentation. Nexarag therefore complements, rather than replaces, existing RAG frameworks by providing an opinionated, researcher-facing system for producing and reusing reproducible knowledge graph contexts in LLM research.
+We chose to build Nexarag rather than contribute directly to an existing RAG or GraphRAG library because our core contribution requires application-level infrastructure that is outside the scope of most libraries’ design constraints: a literature-oriented ingestion workflow (BibTeX and paper lists, citation expansion, and persistent graph storage), a collaborative and visual curation interface, containerized offline deployment, and an MCP-first tool surface for model-agnostic experimentation. Nexarag therefore complements, rather than replaces, existing RAG frameworks by providing a researcher-facing system for producing and reusing reproducible knowledge graph contexts in LLM research.
 
 # Software Design
 In designing Nexarag we focused on four principles: (1) ease-of-use (in deployment and practical application), (2) flexibility (in model selection and configuration), (3) modularity (to promote scale and independent contribution), and (4) privacy and security. For ease-of-use, we chose popular frontend technologies such as Angular, D3.js, and Cytoscape.js and leveraged existing component libraries to provide a simple and familiar frontend user interface with intuitive tools for building knowledge graphs, searching for relevant papers, conversing with LLMs, and visualizing data. Nexarag is also fully containerized, with release artifacts produced by automated build pipelines readily available for local deployment through Docker. For flexibility, we integrate with Ollama and support any embedding model and LLM that is also supported by the user's hardware, making it simple to switch between models for different tasks and as new models are released. Users can also plug in their preferred LLM or coding agent of choice using the built-in MCP server. Additionally, Nexarag features a highly modular design, with distinct services for the REST API, the neo4j knowledge graph, the MCP server, and the frontend application, all bound together with a RabbitMQ messaging backbone. This allows components to scale horizontally, and to minimize the blast radius of contributions in a single service. Finally, Nexarag supports on-premises, air-gapped deployments, offering privacy and security that is not available in cloud-based applications.
