### 2026-01-27 #9568 @yewentao256
Hi @pebabion @gsquared94 @aslan-ng @r0hin would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Nexarag: Democratizing Reproducible Knowledge Graph Contexts for LLM Research".

Thank you so much!
Wentao
