### 2026-01-27 #9568 @yewentao256
Amazing! Usually, once we have two reviewers, we can start the review immediately, but since y'all both responded to quickly, I'll give it another day to see if anyone else bites.

Thank you so much for offering your time, attention, and energy to this JOSS submission!
