### 2026-01-27 #9910 @editorialbot
[REVIEW]: Nexarag: Democratizing Reproducible Knowledge Graph Contexts for LLM Research
**Submitting author:** @tjkerby (<a href="http://orcid.org/0009-0002-1189-3820">Thomas Kerby</a>)
**Repository:** https://github.com/KevinMoonLab/Nexarag
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.0

@pebabion & @aslan-ng & @gsquared94 & @r0hin, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
