### 2026-02-05 #9910 @gsquared94
- Ref: https://github.com/KevinMoonLab/Nexarag/issues/36
   - This project is primarily a systems integration effort, connecting off-the-shelf tools (LangChain, Neo4j, Ollama, Cytoscape) to solve domain-specific problems in literature review.

  > No external engagement.

  - Ref: https://github.com/KevinMoonLab/Nexarag/issues/37
  - No claims to confirm.

   - Ref: https://github.com/KevinMoonLab/Nexarag/issues/38
   - Ref: https://github.com/KevinMoonLab/Nexarag/issues/39
