### 2026-02-08 KevinMoonLab/Nexarag#36 @gsquared94
[JOSS Review] Contribution and authorship
Evaluating this item:
> **Contribution and authorship:** Has the submitting author (@tjkerby) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

While the repository is hosted under `kevinmoonlab`, the project metadata in `src/backend/pyproject.toml` lists only `@tjkerby` as the author, and the git commit history doesn't show any significant commits from @kmoon3.

JOSS [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) state:
> *"Purely financial... and organizational (such as general supervision of a research group) contributions are not considered sufficient... but active project direction and other forms of non-code contributions are."*

@tjkerby @ben-n-fuller Could you please clarify if the co-authors' roles involved **active project direction** or specific non-code contributions (e.g., architecture design, extensive documentation), as opposed to general supervision? This helps confirm the author list meets JOSS criteria.
