### 2026-02-08 KevinMoonLab/Nexarag#38 @gsquared94
[JOSS review] Example usage
It is a bit difficult to verify the interactive workflows solely from the text documentation. Could you please provide a short screencast or video demo demonstrating:
- The Chat UI working end-to-end (uploading a paper -> asking a question).
- The MCP Agent connection (e.g., showing Claude Desktop interacting with the graph).

This would help in verifying the functional claims and would serve as great "Example Usage" for future users.
