### 2026-02-08 KevinMoonLab/Nexarag#39 @gsquared94
[JOSS Review] Automated tests
The project currently lacks automated unit tests for the Python backend, relying instead on manual verification. To ensure long-term maintainability, I strongly recommend adding a standard unit test suite (e.g., pytest) for the backend and enabling a CI workflow to run these tests automatically on Pull Requests.
