### 2026-02-08 KevinMoonLab/Nexarag#40 @gsquared94
[JOSS Review] Community guidelines
- Create a CONTRIBUTING.md file (linking to your existing docs/development/ is a good start).
- Update the README.md to direct bug reports to the GitHub Issues tab.
