+++ docker/prod/mac/docker-compose.yml
@@ -13,7 +13,6 @@ services:
     volumes:
       - neo4j_data:/data
       - ./kg_dumps:/dumps
-      - ./kg_dumps:/var/lib/neo4j/import
     healthcheck:
       test: ["CMD", "cypher-shell", "-u", "neo4j", "-p", "password", "MATCH (n) RETURN COUNT(n);"]
       interval: 10s
@@ -67,7 +66,7 @@ services:
     container_name: nexarag.frontend
     restart: unless-stopped
     ports:
-      - ${FRONTEND_PORT:-5000}:80
+      - ${FRONTEND_PORT:-5100}:80
     depends_on:
       - nexarag.api
 
