### 2026-02-14 #9910 @r0hin
@tjkerby Thank you for the submission! This package is a useful platform for using knowledge graphs to analyse literature. On first pass, here are my suggestions:

1. I did not find standard contribution/support docs, if they exist under nonstandard names/paths, it would be helpful to link them prominently from README and docs.
2. It may be worth creating tagged releases throughout development to follow good practices. Up to you on this though as everything else looks great.

I was able to successfully verify core functionality as described in the README after resolving a few setup problems on my MacOS machine:

3. Address the concerns in my PR https://github.com/KevinMoonLab/Nexarag/pull/41 to improve the installation/setup process via Docker
4. To test functionality, I added a few papers from Semantic Scholar and used the chat function. It would be helpful to add one worked example to the repository that perhaps produces a small graph from a tiny BibTeX input and demonstrates one or two MCP tools, with expected outputs.

And I have a few software paper suggestions:

5. The paper still includes a placeholder string in the repository line: REPLACE-WITH-YOUR-ORG.
6. The Research Impact Statement is credible, but currently light on evidence beyond “near-term potential.” If you could add one or two signals (like benchmarks, integrations, or a reproduced example showing community readiness), it will be much more compelling.
