### 2026-02-18 a5043d0
Update README to include bug reporting link
Added a note to submit bugs to the issues page.
