### 2026-02-20 #9910 @tjkerby
Hi @yewentao256, thanks for the ping, and yes, I’m working through @r0hin’s suggestions.

@r0hin thank you for the detailed review. I agree these will improve the submission. By Feb 25, 2026 I plan to:
- add CONTRIBUTING/support documentation and link it prominently from the README/docs,
- fix the placeholder REPLACE-WITH-YOUR-ORG in the repository line,
- address the Docker-on-macOS improvements in https://github.com/KevinMoonLab/Nexarag/pull/41
- create a tagged release for the JOSS-reviewed version (and can add additional tagged releases going forward).

Separately, I agree the onboarding experience can be improved with additional materials, including a small worked example (tiny BibTeX → small graph + a couple MCP tools with expected outputs), which overlaps with items in https://github.com/KevinMoonLab/Nexarag/issues/38. I’m aiming to complete those by March 6, 2026.
