### 2026-02-25 b85d957
Fix typo and update repository link in paper.md
Corrected a typo in the description of Nexarag and updated the repository link.
