### 2026-02-25 8cd1e28
Fixed security vulnerabilities identified during review
