+++ docker/dev/docker-compose.macos.yml
@@ -43,7 +43,7 @@ services:
     container_name: nexarag.api
     restart: unless-stopped
     ports:
-      - ${API_PORT:-8000}:8000
+      - "127.0.0.1:${API_PORT:-8000}:8000"
     environment:
       NEO4J_URI: bolt://nexarag.neo4j:7687
       NEO4J_USERNAME: ${NEO4J_USERNAME:-neo4j}
@@ -70,7 +70,7 @@ services:
     container_name: nexarag.frontend
     restart: unless-stopped
     ports:
-      - ${FRONTEND_PORT:-500}:80
+      - "127.0.0.1:${FRONTEND_PORT:-500}:80"
     depends_on:
       - nexarag.api
 
@@ -81,7 +81,7 @@ services:
       context: ../../src/backend/src
       dockerfile: nexarag_mcp/Dockerfile
     ports:
-      - "${MCP_PORT:-9000}:9000"
+      - "127.0.0.1:${MCP_PORT:-9000}:9000"
     environment:
       - NEO4J_URI=bolt://nexarag.neo4j:7687
       - ALLOW_DANGEROUS_REQUESTS=true
