+++ docker/prod/mac/docker-compose.yml
@@ -39,7 +39,7 @@ services:
     container_name: nexarag.api
     restart: unless-stopped
     ports:
-      - ${API_PORT:-8000}:8000
+      - "127.0.0.1:${API_PORT:-8000}:8000"
     environment:
       NEO4J_URI: bolt://nexarag.neo4j:7687
       NEO4J_USERNAME: ${NEO4J_USERNAME:-neo4j}
@@ -66,7 +66,7 @@ services:
     container_name: nexarag.frontend
     restart: unless-stopped
     ports:
-      - ${FRONTEND_PORT:-5100}:80
+      - "127.0.0.1:${FRONTEND_PORT:-5100}:80"
     depends_on:
       - nexarag.api
 
@@ -74,7 +74,7 @@ services:
     image: ghcr.io/kevinmoonlab/nexarag/nexarag.mcp:latest
     container_name: nexarag.mcp
     ports:
-      - "${MCP_PORT:-9000}:9000"
+      - "127.0.0.1:${MCP_PORT:-9000}:9000"
     environment:
       - NEO4J_URI=bolt://nexarag.neo4j:7687
       - ALLOW_DANGEROUS_REQUESTS=true
