### 2026-02-25 3f327f4
Add contributing guidelines to CONTRIBUTING.md
