### 2026-02-25 4d949d3
Add contribution guidelines link to README
