### 2026-02-25 b9110ce
Added automated testing infrastructure
