### 2026-02-25 54805ca
Added docker ignore files and improved caching; added authors to pyproject
