+++ src/backend/src/nexarag_mcp/Dockerfile
@@ -1,9 +1,13 @@
+# syntax=docker/dockerfile:1.7
 FROM python:3.11-slim
 WORKDIR /app
+ENV PIP_DISABLE_PIP_VERSION_CHECK=1
 COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+RUN --mount=type=cache,target=/root/.cache/pip \
+    pip install -r requirements.txt
 COPY . /app
-RUN pip install -e /app
+RUN --mount=type=cache,target=/root/.cache/pip \
+    pip install --no-deps -e /app
 ENV PYTHONUNBUFFERED=1
 ENV PYTHONPATH=/app
 CMD ["python", "nexarag_mcp/main.py"]
