+++ src/backend/src/api/Dockerfile
@@ -1,14 +1,17 @@
+# syntax=docker/dockerfile:1.7
 FROM python:3.11-slim
 
 WORKDIR /app
+ENV PIP_DISABLE_PIP_VERSION_CHECK=1
 
-# Install dependencies
 COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+RUN --mount=type=cache,target=/root/.cache/pip \
+    pip install -r requirements.txt
 
 COPY . /app
 
-RUN pip install -e /app
+RUN --mount=type=cache,target=/root/.cache/pip \
+    pip install --no-deps -e /app
 
 ENV PYTHONPATH=/app
 
