+++ src/backend/src/setup.py
@@ -1,8 +1,0 @@
-# setup.py
-from setuptools import setup, find_packages
-
-setup(
-    name="project",
-    version="0.1",
-    packages=find_packages(),
-)
