+++ src/backend/src/requirements.txt
@@ -1,65 +1,0 @@
-aio-pika==9.5.4
-aiohappyeyeballs==2.4.4
-aiohttp==3.11.11
-aiormq==6.8.1
-aiosignal==1.3.2
-annotated-types==0.7.0
-anyio==4.8.0
-attrs==25.1.0
-certifi==2024.12.14
-charset-normalizer==3.4.1
-click==8.1.8
-dataclasses-json==0.6.7
-fastapi==0.115.7
-frozenlist==1.5.0
-greenlet==3.1.1
-h11==0.14.0
-httpcore==1.0.7
-httpx==0.28.1
-idna==3.10
-jsonpatch==1.33
-jsonpointer==3.0.0
-langchain==0.3.15
-langchain-core==0.3.31
-langchain-ollama
-langchain-community
-langchain-neo4j==0.3.0
-langchain-text-splitters==0.3.5
-langsmith==0.3.1
-matplotlib
-multidict==6.1.0
-neo4j==5.26.0
-orjson==3.10.15
-pamqp==3.3.0
-pandas==2.2.0
-propcache==0.2.1
-pydantic==2.10.6
-pydantic_core==2.27.2
-PyMuPDF==1.25.2
-pymupdf4llm==0.0.17
-python-multipart==0.0.20
-pytz==2024.2
-PyYAML==6.0.2
-requests==2.32.3
-requests-toolbelt==1.0.0
-setuptools==75.1.0
-sniffio==1.3.1
-scikit-learn
-SQLAlchemy==2.0.37
-starlette==0.45.3
-tenacity==9.0.0
-torch
-tqdm==4.67.1
-tzdata==2025.1
-urllib3==2.3.0
-uvicorn==0.34.0
-wheel==0.44.0
-yarl==1.18.3
-zstandard==0.23.0
-neomodel==5.4.0
-websockets==15.0
-bibtexparser
-httpx==0.28.1
-fastapi==0.115.7
-mcp[server]
-mcp[cli]
