+++ src/backend/src/nexarag_mcp/Dockerfile
@@ -2,9 +2,10 @@
 FROM python:3.11-slim
 WORKDIR /app
 ENV PIP_DISABLE_PIP_VERSION_CHECK=1
-COPY requirements.txt .
+COPY pyproject.toml /tmp/pyproject.toml
 RUN --mount=type=cache,target=/root/.cache/pip \
-    pip install -r requirements.txt
+    python -c "import tomllib; from pathlib import Path; p=Path('/tmp/pyproject.toml'); d=tomllib.loads(p.read_text())['project']['dependencies']; Path('/tmp/requirements-runtime.txt').write_text('\\n'.join(d)+'\\n')" \
+    && pip install -r /tmp/requirements-runtime.txt
 COPY . /app
 RUN --mount=type=cache,target=/root/.cache/pip \
     pip install --no-deps -e /app
