+++ .github/workflows/backend-tests.yml
@@ -31,7 +31,7 @@ jobs:
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements-test.txt
+          pip install -e "./src[test]"
 
       - name: Run backend unit tests
         env:
