### 2026-02-25 93433ec
Migrated to a single pyproject file; removed requirements
