### 2026-02-25 d5f4bfd
Fixed kg dependencies; moved to uv for dependency installation
