+++ src/backend/pyproject.toml
@@ -7,6 +7,11 @@ name = "nexarag-backend"
 version = "0.1.0"
 description = "NexaRAG backend services and notebook tooling"
 requires-python = ">=3.11"
+authors = [
+  { name = "Tom Kerby", email = "tkerby@stat.byu.edu" },
+  { name = "Ben Fuller", email = "ben.fuller@conexusinc.net" },
+  { name = "Kevin Moon", email = "kevin.moon@usu.edu" },
+]
 dependencies = [
   "aio-pika==9.5.4",
   "bibtexparser",
