### 2026-02-25 6b54045
Moved to uv lock file for python dependency resolution
