+++ .github/workflows/backend-tests.yml
@@ -23,17 +23,17 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      - name: Set up Python
-        uses: actions/setup-python@v5
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v4
         with:
+          version: "0.7.0"
           python-version: '3.11'
 
       - name: Install test dependencies
         run: |
-          python -m pip install --upgrade pip
-          pip install -e ".[test]"
+          uv sync --frozen --extra test
 
       - name: Run backend unit tests
         env:
           PYTHONPATH: src
-        run: pytest
+        run: uv run pytest
