### 2026-02-25 fe159f0
Update feedback section in README.md
Revised feedback section to clarify submission guidelines for bugs, feature requests, and private reports.
