### 2026-02-25 KevinMoonLab/Nexarag#36 @tjkerby
@kmoon3 provided active project direction beyond general supervision; we will update pyproject.toml metadata accordingly.
