### 2026-02-25 #9910 @tjkerby
Hi @r0hin, quick update with links to each item:

- Contributing/support documentation: added and tracked in https://github.com/KevinMoonLab/Nexarag/issues/40
- Placeholder REPLACE-WITH-YOUR-ORG fixed: https://github.com/KevinMoonLab/Nexarag/commit/b85d957eb4c7645fbbc45c580ba5beff1961da0a
- Docker-on-macOS improvements: implemented in https://github.com/KevinMoonLab/Nexarag/pull/42
- Tagged release for the JOSS-reviewed version: https://github.com/KevinMoonLab/Nexarag/releases/tag/v1.0.0

We’re still working on the onboarding improvements (small worked example and expected outputs, overlapping with [#38](https://github.com/KevinMoonLab/Nexarag/issues/38)) and plan to complete those by March 6, 2026.
