### 2026-02-26 #9910 @r0hin
> Hi [@r0hin](https://github.com/r0hin), quick update with links to each item:
>
> * Contributing/support documentation: added and tracked in [[JOSS Review] Community guidelines KevinMoonLab/Nexarag#40](https://github.com/KevinMoonLab/Nexarag/issues/40)
> * Placeholder REPLACE-WITH-YOUR-ORG fixed: [KevinMoonLab/Nexarag@b85d957](https://github.com/KevinMoonLab/Nexarag/commit/b85d957eb4c7645fbbc45c580ba5beff1961da0a)
> * Docker-on-macOS improvements: implemented in [Fixed security vulnerabilities identified during review KevinMoonLab/Nexarag#42](https://github.com/KevinMoonLab/Nexarag/pull/42)
> * Tagged release for the JOSS-reviewed version: https://github.com/KevinMoonLab/Nexarag/releases/tag/v1.0.0
>
> We’re still working on the onboarding improvements (small worked example and expected outputs, overlapping with [#38](https://github.com/KevinMoonLab/Nexarag/issues/38)) and plan to complete those by March 6, 2026.

Looks so far good to me - I've updated my checklist.
