### 2026-03-07 KevinMoonLab/Nexarag#39 @ben-n-fuller
We've added [automated test coverage](https://github.com/KevinMoonLab/Nexarag/tree/main/src/backend/tests) for some of the core backend workflows. These are now triggered automatically for PR validation via [GitHub actions](https://github.com/KevinMoonLab/Nexarag/actions/workflows/backend-tests.yml).
