### 2026-03-07 KevinMoonLab/Nexarag#38 @ben-n-fuller
Nexarag now has a [dedicated YouTube channel](https://www.youtube.com/@Nexarag) with onboarding tutorials for MacOS and Windows deployment, GPU configuration, and MCP integration. These are linked in the repo README.
