### 2026-03-07 KevinMoonLab/Nexarag#37 @ben-n-fuller
These security concerns were addressed in PR #42
