### 2026-03-07 #9910 @tjkerby
@yewentao256, @gsquared94, and @r0hin, here is a quick status update on the remaining onboarding items and reviewer feedback.

**Onboarding materials added**

We created a [YouTube channel](https://www.youtube.com/@Nexarag) and added two walkthrough videos covering (1) installation/setup and (2) interacting with NexaRAG via the app and via MCP (Claude Code + ollmcp). These are linked from the [README/docs](https://github.com/KevinMoonLab/Nexarag).

**Reviewer items addressed**

We’ve now worked through the set of issues raised in the onboarding thread and by @gsquared94: https://github.com/KevinMoonLab/Nexarag/issues/40, https://github.com/KevinMoonLab/Nexarag/issues/39, https://github.com/KevinMoonLab/Nexarag/issues/38, https://github.com/KevinMoonLab/Nexarag/issues/37, and https://github.com/KevinMoonLab/Nexarag/issues/36.

If @gsquared94 is satisfied with the resolutions, we can close those issues. At this point, I believe we have addressed all reviewer feedback raised so far. Please let us know if anything is missing. Otherwise, we believe the review is ready to proceed to the next step.
