### 2026-03-07 #9910 @yewentao256
Thanks @tjkerby for your update, LGTM so far.

For the next step, we need at least 2 explict approval from reviewers (and all checklists marked done), thanks @r0hin @gsquared94 for your contributions to JOSS!

Also happy to see if @pebabion and @aslan-ng could take a look if you are available!
