### 2026-03-09 #9910 @pebabion
@tjkerby  some additional comments:

**Paper**
- Placeholder still present on `joss` branch: The repository link on line 77 of `paper.md` still reads REPLACE-WITH-YOUR-ORG. I see this was fixed on main ([b85d957](https://github.com/KevinMoonLab/Nexarag/commit/b85d957eb4c7645fbbc45c580ba5beff1961da0a)), but the `joss` branch hasn't been synced yet.
-  Missing DOI: The Sentence-BERT reference (reimers-2019-sentence-bert) should include DOI 10.18653/v1/d19-1410, as flagged by editorialbot above
- Paper length: At ~1,812 words the paper is on the long side for JOSS. The Summary and Statement of Need sections overlap significantly so if you want, you can consider tightening this!
- Software Design section: The checklist asks whether this section "explains the trade-offs weighed, the design/architecture chosen, and why it matters." The current text reads more as a feature list (e.g., "we chose popular frontend technologies") than an analysis of design trade-offs considered and rejected. Consider briefly discussing alternatives you evaluated and why they were insufficient.

**Development practices**

- No external engagement: The project has no contributors, issues, or PRs from outside the author team prior to this JOSS review (the one external contribution is @r0hin's review-related PR #41). While the internal development history is solid (~13 months, 249 commits, 22 PRs), external adoption signals are minimal (2 stars, 2 forks). However this is an active development, so it's totally understandable.
- Tests and CI: I see automated tests were recently added on main (backend-tests.yml and 6 pytest files). These cover API sockets/upload/util, RabbitMQ events, and Scholar util - this is a good start! Frontend tests are configured (Jest) but no test files exist yet.

The joss branch appears out of sync with main in several ways (missing CONTRIBUTING.md, missing tests, placeholder in paper), you might want to consider merging them.
