### 2026-03-12 #9910 @gsquared94
> Thanks [@tjkerby](https://github.com/tjkerby) for your update, LGTM so far.
>
> For the next step, we need at least 2 explict approval from reviewers (and all checklists marked done), thanks [@r0hin](https://github.com/r0hin) [@gsquared94](https://github.com/gsquared94) for your contributions to JOSS!
>
> Also happy to see if [@pebabion](https://github.com/pebabion) and [@aslan-ng](https://github.com/aslan-ng) could take a look if you are available!

Marked checklist items as done, and approved from my end.
