### 2026-03-13 #9910 @tjkerby
@pebabion, thank you for the thorough review and for catching the joss branch drift.

Planned updates (target: Monday, March 16, 2026), with links once merged:

**Paper**
- Sync joss with main so fixes already on main are reflected (including the REPLACE-WITH-YOUR-ORG repository link in paper.md, line 77).
- Add the missing DOI for the Sentence-BERT reference (DOI: 10.18653/v1/d19-1410).
- Tighten the paper by reducing overlap between Summary and Statement of Need.
- Revise the Software Design section to focus on design trade-offs, including a brief note on alternatives we considered and why we did not choose them (instead of listing features/technologies).

**Development practices**
- Add at least a minimal set of frontend tests to match the existing Jest configuration.
- Ensure the joss branch includes the current CONTRIBUTING docs and the recently added tests/CI from main.

I’ll post a follow-up comment here with the PR(s)/commit(s) for each item as soon as they are in place.
