### 2026-03-16 b3cd4e7
Tightened paper language and added tradeoffs for design choices; added frontend unit test coverage
