+++ .github/workflows/frontend-tests.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Set up Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: '22'
+          node-version: '23'
           cache: 'npm'
           cache-dependency-path: src/frontend/package-lock.json
 
