+++ paper/paper.bib
@@ -162,7 +162,8 @@
   month     = {November},
   year      = {2019},
   publisher = {Association for Computational Linguistics},
-  url       = {http://arxiv.org/abs/1908.10084}
+  url       = {http://arxiv.org/abs/1908.10084},
+  doi       = {10.18653/v1/d19-1410}
 }
 
 @article{reinanda2020knowledge,
