### 2026-03-16 #9910 @tjkerby
@pebabion and @yewentao256, thank you, we’ve implemented the requested updates.

**Paper + code updates (PR):** [https://github.com/KevinMoonLab/Nexarag/pull/45](https://github.com/KevinMoonLab/Nexarag/pull/45)

* Tightened the paper by reducing overlap between **Summary** and **Statement of Need** (overall word count reduced).
* Revised the **Software Design** section.
* Added initial frontend tests to match the existing Jest setup.

**Reference fix (commit):** [https://github.com/KevinMoonLab/Nexarag/commit/35c24b530ca41fd4d68879ce2eedb22240d0b624](https://github.com/KevinMoonLab/Nexarag/commit/35c24b530ca41fd4d68879ce2eedb22240d0b624)

* Added the missing DOI for the Sentence-BERT reference (`10.18653/v1/d19-1410`).

**Branch sync**

* We synced `joss` with `main` so the previously fixed placeholder and other recent additions (e.g., CONTRIBUTING/tests) are now present on the `joss` branch as well.

Please let us know if anything is still missing or if you’d like further adjustments.
