### 2026-03-30 #9910 @pebabion
> [@pebabion](https://github.com/pebabion) thank you again for the careful review. It looks to me that the items from your March 8th comment have now been addressed.
>
> If you have time for a final pass, I’d really appreciate a quick confirmation here. Otherwise, to keep the review moving, I plan to proceed after April 1, 2026 based on the two completed reviewer approvals.

Everything looks good from my end. I've updated my checklist accordingly.
