### 2026-03-30 #9910 @yewentao256
Hi @tjkerby , some issues I found:
- Duplicate URLs: like Anthropic and a lot of other places

- Wrong ref for OPENAI
- `Kinney et al. (2023)` is not the standard DOI

Would recommend take a thorough reading, smooth a few remaining awkward phrases in the main text and headings
