### 2026-04-01 dab4a7a
Revise software design principles section in paper
