+++ paper/paper.md
@@ -47,12 +47,12 @@ Literature discovery platforms such as Connected Papers, ResearchRabbit, and Lit
 # Software Design
 Nexarag was designed around four principles: ease of use, flexibility, modularity, and privacy/security.
 
-| Principle | Design choices in Nexarag | Tradeoffs |
-| - | - | - |
-| **Ease of use** | Nexarag uses familiar frontend technologies, including Angular, D3.js, Cytoscape.js, and PrimeNg, to provide an intuitive interface for building knowledge graphs, finding papers, interacting with LLMs, and visualizing results. Nexarag is fully containerized and can be deployed with a single command. | Multiple component libraries increases frontend maintenance overhead and onboarding cost for new contributors. Container deployments add complexity and additional software dependencies (Docker). |
-| **Flexibility** | Nexarag integrates with Ollama and supports any embedding model or LLM that the user’s hardware can run, making it easy to switch models across tasks or adopt new ones as they become available. Users can also connect their preferred LLM or coding agent through the built-in MCP server. | Users have more choices but also more responsibility for hardware configuration, model selection, and staying up-to-date on relevant tools and architectures. |
-| **Modularity** | The system is organized as distinct services for the REST API, Neo4j knowledge graph, MCP server, and frontend application, connected through a RabbitMQ messaging backbone. This supports horizontal scaling and reduces the blast radius of changes made within any single service. | Service decomposition improves scalability and isolates changes, but increases deployment complexity, inter-service coordination, and operational overhead. |
-| **Privacy and security** | Nexarag supports on-premises, air-gapped deployment, providing a level of privacy and security that cloud-based applications typically cannot offer. | Air-gapped deployments can offer heightened security and privacy, but place more burden on the user for hardware configuration, deployment, and maintenance. Local compute resources may also be limited compared to cloud services. |
+| Design choices in Nexarag | Tradeoffs |
+| - | - |
+|  **Ease of use**: Nexarag uses familiar frontend technologies, including Angular, D3.js, Cytoscape.js, and PrimeNg, to provide an intuitive interface for building knowledge graphs, finding papers, interacting with LLMs, and visualizing results. Nexarag is fully containerized and can be deployed with a single command. | Multiple component libraries increases frontend maintenance overhead and onboarding cost for new contributors. Container deployments add complexity and additional software dependencies (Docker). |
+| **Flexibility**: Nexarag integrates with Ollama and supports any embedding model or LLM that the user’s hardware can run, making it easy to switch models across tasks or adopt new ones as they become available. Users can also connect their preferred LLM or coding agent through the built-in MCP server. | Users have more choices but also more responsibility for hardware configuration, model selection, and staying up-to-date on relevant tools and architectures. |
+| **Modularity**: The system is organized as distinct services for the REST API, Neo4j knowledge graph, MCP server, and frontend application, connected through a RabbitMQ messaging backbone. This supports horizontal scaling and reduces the blast radius of changes made within any single service. | Service decomposition improves scalability and isolates changes, but increases deployment complexity, inter-service coordination, and operational overhead. |
+| **Privacy and security**: Nexarag supports on-premises, air-gapped deployment, providing a level of privacy and security that cloud-based applications typically cannot offer. | Air-gapped deployments can offer heightened security and privacy, but place more burden on the user for hardware configuration, deployment, and maintenance. Local compute resources may also be limited compared to cloud services. |
 
 
 # Software overview
