### 2026-04-08 6c1a70e
Modify author affiliations and section titles
Updated author affiliations and section headers in the paper.
