+++ paper/paper.md
@@ -10,15 +10,15 @@ authors:
   - name: Thomas J. Kerby
     affiliation: 1
   - name: Benjamin N. Fuller
-    affiliation: 3
-  - name: Kevin R. Moon
     affiliation: 2
+  - name: Kevin R. Moon
+    affiliation: 3
 affiliations:
   - name: Brigham Young University, Provo, UT
     index: 1
-  - name: Utah State University, Logan, UT
-    index: 2
   - name: Independent Researcher
+    index: 2
+  - name: Utah State University, Logan, UT
     index: 3
 date: 2025-11-04
 bibliography: paper.bib
@@ -29,13 +29,13 @@ bibliography: paper.bib
 Large language models (LLMs) are widely used in research workflows but struggle with hallucinations, short context windows, and weak reproducibility in literature reviews [@Ji2023; @Huang2025]. Nexarag is a modular, open‑source platform that lets researchers curate, visualize, and share custom knowledge graphs (KGs) from academic sources stored in Neo4j [@neo4j2024database]. Through native support for the Model Context Protocol (MCP), any MCP‑compatible LLM can access these curated KGs for controllable, reproducible context injection [@anthropic2024mcp; @mcp2024github]—including fully private, air‑gapped deployments via containers [@boettiger2015docker]—so teams can explore literature more deeply and transparently. Nexarag provides interactive graph/semantic visualizations using Cytoscape.js and D3 [@franz2023cytoscape; @bostock2011d3].
 
 
-## Statement of need
+# Statement of need
 
 Retrieval-augmented generation (RAG) has become a standard approach for knowledge-intensive NLP [@Lewis2020; @Guu2020; @gao2023retrieval], but systems built primarily on embedding-based similarity [@Lewis2020; @Guu2020; @reimers-2019-sentence-bert] can miss long-range semantic structure and cross-document relationships, especially in long-context and multi-document settings [@wang-etal-2024-leave; @gao2023retrieval]. For literature synthesis and related research workflows, this limits transparency, controllability, and reproducibility.
 
 Knowledge graphs address part of this problem by representing entities and relations explicitly, enabling path-based queries and more interpretable reasoning over document collections [@reinanda2020knowledge; @sahlab2022knowledge; @Xu2024]. However, existing KG-based tooling is often either proprietary or too technically demanding for routine research use. Nexarag addresses this gap with a researcher-friendly, self-hostable platform for constructing and curating literature knowledge graphs and exposing them to language models through MCP [@anthropic2024mcp].
 
-## State of the field
+# State of the field
 
 Open-source RAG frameworks such as LangChain, LlamaIndex, and Haystack provide reusable components for ingestion, indexing, and retrieve-then-generate pipelines [@langchain2026rag; @llamaindex2026rag; @haystack2026]. They are effective developer libraries, but they are mainly designed for assembling application-specific pipelines in code.
 
@@ -44,8 +44,9 @@ Graph-augmented retrieval tools extend this ecosystem. LlamaIndex includes a Kno
 Literature discovery platforms such as Connected Papers, ResearchRabbit, and Litmaps support citation exploration and related-work discovery [@connectedpapers2026; @researchrabbit2026; @litmaps2026], but they do not provide a researcher-owned, versionable graph substrate for controlled LLM experiments. Nexarag sits between these tool families by packaging persistent Neo4j-based graph construction, interactive curation, and standardized model access through MCP into a self-hostable research application [@anthropic2024mcp; @mcp2024github]. It complements existing RAG and GraphRAG libraries by turning graph-based context construction into a reusable and shareable research workflow with user-friendly UI tools, visualization, and pluggable MCP integration.
 
 
-# Software Design
-Nexarag was designed around four principles: ease of use, flexibility, modularity, and privacy/security.
+# Software design
[…]
