### 2026-04-08 6356f13
Document Nexarag features and design choices
Added features and capabilities of Nexarag research platform.
