### 2026-04-08 #9910 @tjkerby
Hi @yewentao256, thank you for the careful review. We have revised the manuscript and bibliography in response to your comments (see https://github.com/KevinMoonLab/Nexarag/pull/46). The changes include simplifying the paper structure to better match the expected JOSS format, removing redundant subsections and README-style material, tightening the prose, and cleaning up duplicated content in the AI usage disclosure. We also corrected the bibliography issues you noted, including duplicate URLs, the OpenAI reference, and the Kinney et al. (2023) entry. Please let me know if there are any other changes you would like to see.
