### 2026-04-10 #9910 @yewentao256
Hi @tjkerby Thank you for the updates. I think we are now close to wrapping up the post-review process.

Before I proceed, could you please post:

- the final software version number to be used for the JOSS paper;
- the archive DOI for the archived release (e.g., Zenodo/figshare); and
- a brief confirmation that the archive title, author list (including ORCIDs, where applicable), and license match those in the JOSS paper.

See "Additional Author Tasks After Review is Complete" above
