### 2026-04-15 #9910 @tjkerby
Hi @yewentao256, that is fantastic news. Here is the requested information:
- Software version 1.1.0
- DOI: https://doi.org/10.5281/zenodo.19595058
- We confirm that the title, author list, and license match those in the joss paper.

Let us know if you need anything else.
