### 2026-04-22 #9910 @crvernon
Editor's Note:  Though @pebabion has open items in their checklist, I am considering these addressed per [this comment](https://github.com/openjournals/joss-reviews/issues/9910#issuecomment-4151853193).

@tjkerby - this submission is looking great.  I just need you to address the following capitalization issues in your references via your bib file before I accept.  Note:  you can maintain formatting of characters in your bib file by wrapping those characters in curly brackets.

- [x] LINE 105:  The "d" in "docker" should be capitalized.
- [x] LINE 111:  The "h" in "(haystack..." should be capitalized.
- [x] LINE 116:  Both occurrences of the "p" in "python" should be capitalized.
- [x] LINE 154:  The "l" in "linux" should be capitalized.
- [x] LINE 161:  The "p" in "python" should be capiatalized.
- [x] LINES 165-166:  Should be capitalized as "Get up and running with Llama 3.2, Mistral, Gemma 2..."

Let me know when these changes have been made and, after confirmation, I'll accept.
