### 2026-04-24 fd89dcb
Format specific terms in bibliography titles
Updated titles in the bibliography to format specific terms in braces for consistency.
