+++ paper/paper.bib
@@ -209,7 +209,7 @@
 }
 
 @article{boettiger2015docker,
-  title = {An introduction to Docker for reproducible research},
+  title = {An introduction to {Docker} for reproducible research},
   author = {Boettiger, Carl},
   journal = {ACM SIGOPS Operating Systems Review},
   volume = {49},
@@ -230,7 +230,7 @@
 }
 
 @misc{fastapi2024,
-  title = {FastAPI: Modern, fast (high-performance), web framework for building APIs with Python 3.6+ based on standard Python type hints},
+  title = {FastAPI: Modern, fast (high-performance), web framework for building APIs with {Python} 3.6+ based on standard {Python} type hints},
   author = {{FastAPI Contributors}},
   year = {2024},
   url = {https://fastapi.tiangolo.com/},
@@ -246,7 +246,7 @@
 }
 
 @misc{ollama2024,
-  title = {Ollama: Get up and running with Llama 3.2, Mistral, Gemma 2, and other large language models},
+  title = {Ollama: Get up and running with {Llama} 3.2, {Mistral}, {Gemma} 2, and other large language models},
   author = {{Ollama Team}},
   year = {2024},
   url = {https://ollama.com/},
@@ -284,7 +284,7 @@
 }
 
 @article{merkel2014docker,
-  title = {Docker: lightweight linux containers for consistent development and deployment},
+  title = {Docker: lightweight {Linux} containers for consistent development and deployment},
   author = {Merkel, Dirk},
   journal = {Linux Journal},
   volume = {2014},
@@ -310,7 +310,7 @@
 }
 
 @misc{haystack2026,
-  title = {Get Started (Haystack Documentation)},
+  title = {Get Started ({Haystack} Documentation)},
   author = {{deepset Haystack}},
   year = {2026},
   url = {https://docs.haystack.deepset.ai/docs/get-started},
@@ -342,7 +342,7 @@
 }
 
 @misc{neo4j2026graphrag,
-  title = {neo4j/neo4j-graphrag-python: Neo4j GraphRAG Package for Python},
+  title = {neo4j/neo4j-graphrag-python: Neo4j GraphRAG Package for {Python}},
   author = {{Neo4j}},
   year = {2026},
   url = {https://github.com/neo4j/neo4j-graphrag-python},
