### 2026-04-24 #9910 @tjkerby
Thank you @crvernon for the careful review and for pointing out the remaining capitalization issues. I have now updated the affected entries in `paper.bib`, including the referenced occurrences of Docker, Haystack, Python, Linux, and the Ollama title text, using protected capitalization to preserve the intended formatting.

The changes are included in https://github.com/KevinMoonLab/Nexarag/pull/47. Please let me know if there are any remaining issues to address, but I believe these final reference updates are now complete.
