### 2026-04-25 #9910 @crvernon
🥳 Congratulations on your new publication @tjkerby! Many thanks to @yewentao256 for editing and @pebabion, @gsquared94, and @r0hin for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
