### 2026-04-28 #9910 @pebabion
> 🥳 Congratulations on your new publication [@tjkerby](https://github.com/tjkerby)! Many thanks to [@yewentao256](https://github.com/yewentao256) for editing and [@pebabion](https://github.com/pebabion), [@gsquared94](https://github.com/gsquared94), and [@r0hin](https://github.com/r0hin) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
>
> Please consider becoming a reviewer for JOSS if you are not already: https://reviewers.joss.theoj.org/join

Congratulations!
